Added Classes for future use and adjusted player movement
Player now moves from tile to tile.
GameScreen class is for future use

+++ src/Player.h
@@ -44,7 +44,7 @@ namespace Game {
 
         Vector2 getPos();
 
-        void setPos(int inputX, int inputY);
+        void setPos(float inputX, float inputY);
 
         Texture2D getTexture();
